backend/requirements.txt
Django==4.2.28
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==12.1.1
drf-spectacular==0.26.5
django-filter==23.3
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==22.0.0
django-extensions==3.2.3
coverage==7.3.2
